DebugForVectorValuesAfterMahiComment
Actually we are generating vectors values that will be 3072 array size, but this will be passed to cosine similarity, as collection[0] which means internally it has all vector values, It quite confusing to map this vector value among compared similarity.. we need to improve our understanding in this. as Cosine similarity generates a single by comparing all these vector values between two docs.